multiagent_interaction/requirements.txt
# Core AI/ML frameworks
langchain>=0.1.0
langgraph>=0.0.40
openai>=1.0.0
anthropic>=0.18.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.3.0

# Multi-agent frameworks
autogen-agentchat>=0.2.0
pyautogen>=0.2.0

# Experiment design
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
itertools  # built-in

# Causal inference and debiasing
dowhy>=0.11.0
econml>=0.14.0
causalnex>=0.12.0

# Data handling
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.1

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
